Create UsernameSearch component which holds validation for username; use it in two places UserSettings and SetUsername;
