+++ src/components/editForm/user.js
@@ -1,13 +1,11 @@
 // @flow
 import * as React from 'react';
 import { withRouter } from 'react-router';
-import slugg from 'slugg';
 import { withApollo } from 'react-apollo';
 import compose from 'recompose/compose';
 import { connect } from 'react-redux';
 import Link from 'src/components/link';
 import { track } from '../../helpers/events';
-import { throttle } from '../../helpers/utils';
 import { Button } from '../buttons';
 import Icon from '../../components/icons';
 import { SERVER_URL, CLIENT_URL } from '../../api/constants';
@@ -20,6 +18,7 @@ import {
   PhotoInput,
   CoverInput,
 } from '../formElements';
+import UsernameSearch from '../usernameSearch';
 import { StyledLabel } from '../formElements/style';
 import {
   StyledCard,
@@ -28,12 +27,8 @@ import {
   Actions,
   ImageInputWrapper,
   Location,
-  Loading,
   GithubSignin,
 } from './style';
-import { Spinner } from '../../components/globals';
-import { getUserByUsernameQuery } from 'shared/graphql/queries/user/getUser';
-import type { GetUserType } from 'shared/graphql/queries/user/getUser';
 import editUserMutation from 'shared/graphql/mutations/user/editUser';
 import type { EditUserType } from 'shared/graphql/mutations/user/editUser';
 import { addToastWithTimeout } from '../../actions/toasts';
@@ -61,7 +56,6 @@ type State = {
   photoSizeError: string,
   proGifError: boolean,
   usernameError: string,
-  isUsernameSearching: boolean,
 };
 
 type Props = {
@@ -93,10 +87,7 @@ class UserWithData extends React.Component<Props, State> {
       photoSizeError: '',
       proGifError: false,
       usernameError: '',
-      isUsernameSearching: false,
     };
-
-    this.search = throttle(this.search, 500);
   }
 
   changeName = e => {
@@ -309,80 +300,18 @@ class UserWithData extends React.Component<Props, State> {
       });
   };
 
-  changeUsername = e => {
-    let username = e.target.value.trim();
-    username = slugg(username);
-
+  handleUsernameValidation = ({ error, username }) => {
+    const { currentUser } = this.props;
+    // we want to reset error if was typed same username which was set before
+    const usernameError = currentUser.username === username ? '' : error;
     this.setState({
-      usernameError: '',
+      usernameError,
       username,
     });
-
-    if (username.length > 20) {
-      return this.setState({
-        usernameError: 'Usernames can be up to 20 characters',
-      });
-    } else if (username.length === 0) {
-      this.setState({
-        usernameError: 'Be sure to set a username so that people can find you!',
-      });
-    } else {
-      this.setState({
-        usernameError: '',
-      });
-    }
-
-    // $FlowIssue
-    this.search(username);
   };
 
-  search = (username: string) => {
-    if (username.length > 20) {
-      return this.setState({
-        usernameError: 'Usernames can be up to 20 characters',
-        isUsernameSearching: false,
-      });
-    } else if (username.length === 0) {
-      return this.setState({
-        usernameError: 'Be sure to set a username so that people can find you!',
-        isUsernameSearching: false,
-      });
-    } else {
-      this.setState({
-        usernameError: '',
-        isUsernameSearching: true,
-      });
-
-      // check the db to see if this channel slug exists
-      this.props.client
-        .query({
-          query: getUserByUsernameQuery,
-          variables: {
-            username,
-          },
-        })
-        .then(({ data: { user } }: { data: { user: GetUserType } }) => {
-          if (this.state.username.length > 20) {
-            return this.setState({
-              usernameError: 'Usernames can be up to 20 characters',
-              isUsernameSearching: false,
-            });
-          } else if (user && user.id) {
-            return this.setState({
-              usernameError: 'This username is already taken, sorry!',
-              isUsernameSearching: false,
-            });
-          } else {
-            return this.setState({
-              usernameError: '',
-              isUsernameSearching: false,
-            });
-          }
-        })
-        .catch(err =>
-          this.props.dispatch(addToastWithTimeout('error', err.message))
-        );
-    }
+  handleOnError = err => {
+    this.props.dispatch(addToastWithTimeout('error', err.message));
   };
 
   render() {
@@ -401,7 +330,6 @@ class UserWithData extends React.Component<Props, State> {
       photoSizeError,
       proGifError,
       usernameError,
-      isUsernameSearching,
     } = this.state;
 
     const postAuthRedirectPath = `?r=${CLIENT_URL}/users/${username}/settings`;
@@ -453,18 +381,14 @@ class UserWithData extends React.Component<Props, State> {
 
           {nameError && <Error>Names can be up to 50 characters.</Error>}
 
-          <Input
+          <UsernameSearch
             type={'text'}
-            defaultValue={username}
-            onChange={this.changeUsername}
-          >
-            Username
-            {isUsernameSearching && (
-              <Loading>
-                <Spinner size={16} color={'brand.default'} />
-              </Loading>
-            )}
-          </Input>
+            label="Username"
+            size={'small'}
+            username={username}
+            onValidationResult={this.handleUsernameValidation}
+            onError={this.handleOnError}
+          />
 
           {usernameError && (
             <Notice style={{ marginTop: '16px' }}>{usernameError}</Notice>
